Introduce SerializeTo trait.
SerializeTo is an object safe form of Serialize; by lifting the
serializer parameter to the trait, it is possible to create trait
objects of SerializeTo if the user can specify the Serializer in
advance.

The definition of Serializer changes to take SerializeTo<Self>
objects anywhere  that Serialize is currently taken.

This is a breaking change, so the version is bumped to 0.9.0.