Allow empty_struct to deserialize units, too
Annotating a unit struct or variant with #[serde(empty_struct)] will still cause
it to be serialized to an empty struct, but now the generated deserializer will
accept either a unit or empty struct value. So, for example, a JSON protocol
with an adjacently-tagged enum for messages will accept both

        {"tag": "HelloWorld", content: {}}

and

        {"tag": "HelloWorld"}

if the variant HelloWorld is annotated with #[serde(empty_struct)], but will
default to serializing to the former.