Allow `#[serde(flatten)]` on seq represented structures
This commit allows `#[serde(flatten)]` for tuple structs, and generates
seq deserialization code for named structs containing
`#[serde(flatten)]`. There's no fundamental reason we can't allow this,
and in fact we can do this much more efficiently for sequences than we
can for maps.

With maps we need to collect all the unknown keys, and then pass them to
the flattened structures. For sequences, however, we just need to
deserialize the flattened structure using the same `SeqAccess` that
we're using. Due to the nature of that trait, it will be advanced as
much as needed for the underlying structure.

I've skipped in-place deserialization when this attr is present (similar
to what we already do for named structs), as I'm unfamiliar with the
feature or it's implementation. However, I suspect there's no reason we
can't support it easily, since unlike with maps, we do not need to
collect the flattened values separately.

My use case for this is the need to deserialize a structure in the form:

```
["string_content", (TStringContent | StringEmbexpr | StringDVar)*]
```

There's no way to deserialize this structure with a derive in Serde
prior to this commit. The struct representation would be

```rust
struct StringContent(string_content_tag, Vec<StringContentPart>);
```

The manual deserialize impl for this would look roughly like:

```rust
fn visit_seq<A>(seq: A) -> Result<Self::Item, A::Error> {
    let tag = seq.next_element()?
        .ok_or_else(|| Error::invalid_length(0, &self))?;
    let parts = Deserialize::deserialize(
        SeqAccessDeserializer::new(&mut seq),
    )?;
    Ok(Self(tag, parts))
}
```

With this commit, this struct can now be written as:

```rust
 #[derive(Deserialize)]
struct StringContent(string_content_tag, #[serde(flatten)] Vec<StringContentPart>);
```

or

```rust
 #[derive(Deserialize)]
struct StringContent {
    tag: string_content_tag,
    #[serde(flatten)]
    parts: Vec<StringContentPart>,
}
```

It is worth noting that this can lead to some funky error messages for
any missing fields after a flattened field. Since we don't know how many
values the flattened field consumed, we can't actually say how many
values we got. For example, if you try to deserialize a 3 element array
into `struct Foo(i32, #[serde(flatten)] (i32, i32), i32)`, we'll say we
only got 2 elements in the error. I'm not entirely sure what (if
anything) we want to do about this, since the API of `Error` doesn't
really support "I don't know how many fields I encountered"