Provide serialisation context to data structures
In order for representations in the serde data model to be adaptable to
particular requirements from different serialisation formats, we provide
means for the de/serializer to be queried for additional context.

This is an escape hatch for niche situations, that increases coupling
between the data structure and the de/serializer and should therefore be
avoided wherever possible.  Neither serialisation format nor data
structure should assume/require that the other supports any particular
context, but should instead adopt sensible defaults whenever an expected
context is unsupported.

Closes #2003