Add sequentially_adjacent_tag representation for enums.
If present, this attribute will cause enums to be serialized and
deserialized into a sequence. The first element is the tag, and the
second the content. If the content is optional, then the sequence may
only be a single item.

Also, add a test to verify the functionality of this new feature.