Add Container::struct_as_map()
Sometimes a struct is encoded as a map. This is currently only when the struct
contains a flattened field but will also be the case when non-string names are
supported. Add a predicate to determine whether a struct should be encoded as a
map so the logic can be easily updated.

No functional change.