Add support for integer renamed struct fields
The existing #[serde(rename = ...)] syntax is extended to support interger
names for struct fields. The type of the literal informs the serialization
function that is used for the field name with the default being i32, to match
the Rust's default.

Deserialization is more flexible and accepts any integer type, matching against
the field values as 64-bit integers. This allows better compatibility with
deserializers that don't know the exact bit width of integer values as well as
those that do.