Use fallthough to collect unknown field names
To be consistent across all types, use the fallthough to collect non-string
field names rather than having different logic between the types.

When collect_other_fields is true, it's already assumed that the fallthrough is
going to expect __value to be a variant of Content. That assumption is now
applied all the types, not just the string-like types.