Remove explicit and implied references to specific MSRVs
After conversation in Twitter DM re: #2255, it became clear that
serde does not, in fact, have a specific MSRV policy, but rather
strives to be buildable with older rustc as only a practical
matter.

This led to some confusion from users, as the `rust-version`
Cargo.toml tag is often used to communicate a specific MSRV policy.

Thus, this PR simply removes references which could be confused for
a specific MSRV  policy, leaving a comment in CI that the test with
older rustc's does not imply a specific policy.