Make the serde helper process persistent.
Rather than starting the helper process for every proc-macro invocation,
start it on the first macro invocation and keep it around. The goal is to
amortize the process startup cost across multiple macro invocations.

The process loops waiting for multiple requests, until it gets an
EOF on its stdin, whereupon it exits. This will happen when the rustc
process exits and closes its end of the stdin pipe.

To do this, this change extends the protocol to prefix each chunk
of data with a 32-bit little-endian length, so that each side knows
how much to read, rather than relying on reading to EOF.