Use deserialize_map when deserialize internally tagged enums
Deserializer methods are only hints which deserializer is not obliged to follow.

Both
- TaggedContentVisitor
- InternallyTaggedUnitVisitor

accepts only visit_map and visit_seq and that is what derived implementation of
Deserialize does for structs. Therefore it is fine to call deserialize_map here,
as that already did in derived deserialize implementation. Because those structs
officially not public, it is used only by derive macro