serde_derive: Add alias_all attribute to containers
This adds a new attribute `#[serde(alias_all = "...")]` to containers,
which accepts the same rename rules as `#[serde(rename_all = "...")]`.
When applied to a container, the alias will be added to all fields (or variants)
similarly to how `#[serde(alias = "...")]` works on a single field/variant.

This attribute can be specified multiple times, with the use-case being
accepting multiple case styles during de-serialization.

Resolves #1530

Signed-off-by: Michel Heily <michelheily@gmail.com>